chore: Unvendor pydantic_argparse and wire up uv correctly
