+++ src/gallia/plugins/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: AISEC Pentesting Team
+#
+# SPDX-License-Identifier: Apache-2.0
+
