+++ src/gallia/power_supply/exceptions.py
