+++ src/gallia/pydantic_argparse/parsers/enum.py
@@ -14,7 +14,7 @@ import argparse
 import enum
 from typing import Any
 
-from pydantic_argparse.utils.pydantic import PydanticField
+from gallia.pydantic_argparse.utils.pydantic import PydanticField
 
 from .utils import SupportsAddArgument
 from ..utils.field import ArgFieldInfo
